# public/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Anime API",
    "description": "RESTful API for anime information, streaming, and discovery. Built with Node.js and Express.\n\n**Disclaimer:** This API does not store any files. It only links to media hosted on third-party services. Made for educational purposes only.\n\n**Provider System:** All endpoints support a `provider` query parameter to select the data source. Available providers: `9anime` (default, 9animetv.to), `kaido` (kaido.to). If a provider fails, the system automatically falls back to the next available provider.\n\nAll endpoints return responses wrapped in:\n```json\n{ \"success\": true, \"results\": { ... } }\n```\nOn error:\n```json\n{ \"success\": false, \"message\": \"error description\" }\n```",
    "version": "1.0.0",
    "license": {
      "name": "MIT",
      "url": "https://opensource.org/licenses/MIT"
    },
    "contact": {
      "name": "GitHub Repository",
      "url": "https://github.com/kacalayar/aniapi"
    }
  },
  "servers": [
    {
      "url": "/",
      "description": "Current server"
    }
  ],
...
  "paths": {
...
    "/api/info": {
      "get": {
        "tags": ["Anime"],
        "summary": "Anime details",
        "description": "Returns full details for a specific anime including metadata, synopsis, seasons, characters, and recommendations.",
        "operationId": "getAnimeInfo",
        "parameters": [
          {
            "name": "id",
            "in": "query",
            "required": true,
            "description": "Anime slug identifier (e.g. `frieren-beyond-journeys-end-18542`)",
            "schema": { "type": "string" },
            "example": "frieren-beyond-journeys-end-18542"
          },
          { "$ref": "#/components/parameters/ProviderParam" }
        ],
        "responses": {
          "200": {
            "description": "Anime details",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": { "type": "boolean", "example": true },
                    "results": {
                      "type": "object",
                      "properties": {
                        "data": { "$ref": "#/components/schemas/AnimeDetail" },
                        "seasons": {
                          "type": "array",
                          "items": { "$ref": "#/components/schemas/Season" }
                        }
                      }
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/api/watchlist/{userId}/{page}": {
      "get": {
        "tags": ["Watchlist"],
        "summary": "User watchlist",
        "description": "Returns paginated watchlist for a specific user.",
        "operationId": "getWatchlist",
        "parameters": [
          {
            "name": "userId",
            "in": "path",
            "required": true,
            "description": "User identifier",
            "schema": { "type": "string" }
          },
          {
            "name": "page",
            "in": "path",
            "required": false,
            "description": "Page number (default: 1)",
            "schema": { "type": "integer", "default": 1 }
          },
          { "$ref": "#/components/parameters/ProviderParam" }
        ],
        "responses": {
          "200": {
            "description": "Watchlist data",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": { "type": "boolean", "example": true },
                    "results": { "$ref": "#/components/schemas/PaginatedAnimeList" }
                  }
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "parameters": {
      "ProviderParam": {
        "name": "provider",
        "in": "query",
        "required": false,
        "description": "Data source provider. If the selected provider fails, the system automatically falls back to the next available one.",
        "schema": {
          "type": "string",
          "enum": ["9anime", "kaido"],
          "default": "9anime"
        },
        "example": "9anime"
      }
    },
    "schemas": {
...
      "AnimeCard": {
        "type": "object",
        "description": "Anime card from category/browse pages",
        "properties": {
          "id": { "type": "string" },
          "data_id": { "type": "string" },
          "poster": { "type": "string" },
          "title": { "type": "string" },
          "japanese_title": { "type": "string" },
          "description": { "type": "string" },
          "tvInfo": {
            "type": "object",
            "properties": {
              "showType": { "type": "string", "description": "HD, Unknown, or enriched type from qtip" },
              "sub": { "type": ["integer", "boolean", "null"], "description": "Episode count if available, true if present, null if absent" },
              "dub": { "type": ["integer", "boolean", "null"], "description": "Episode count if available, true if present, null if absent" },
              "eps": { "type": "integer", "nullable": true, "description": "Total episode count" }
            }
          },
          "adultContent": { "type": "boolean" }
        }
      },
...
      "AnimeDetail": {
        "type": "object",
        "properties": {
          "adultContent": { "type": "boolean" },
          "id": { "type": "string" },
          "data_id": { "type": "integer" },
          "title": { "type": "string" },
          "japanese_title": { "type": "string" },
          "poster": { "type": "string" },
          "showType": { "type": "string" },
          "animeInfo": {
            "type": "object",
            "properties": {
              "Overview": { "type": "string" },
              "Japanese": { "type": "string" },
              "Synonyms": { "type": "string" },
              "Aired": { "type": "string" },
              "Premiered": { "type": "string" },
              "Duration": { "type": "string" },
              "Status": { "type": "string" },
              "MAL Score": { "type": "string" },
              "Genre": { "type": "array", "items": { "type": "string" } },
              "Studios": { "type": "string" }
            }
          }
        }
      },
      "Season": {
        "type": "object",
        "properties": {
          "id": { "type": "string" },
          "data_number": { "type": "integer" },
          "data_id": { "type": "integer" },
          "season": { "type": "string" },
          "title": { "type": "string" },
          "japanese_title": { "type": "string" },
          "season_poster": { "type": "string" }
        }
      },
...
      "PaginatedAnimeList": {
        "type": "object",
        "properties": {
          "totalPages": { "type": "integer" },
          "data": {
            "type": "array",
            "items": { "$ref": "#/components/schemas/AnimeCard" }
          }
        }
      },
...
    }
  }
}